# contract/openapi.yaml
openapi: 3.1.0
info:
  title: КОСМОБЕРЕГ API
  version: 1.0.0
  description: |
    Контракт между фронтендом и бэкендом платформы КОСМОБЕРЕГ.

    ## Как этим пользоваться

    Это **единственный источник истины** об интерфейсе между фронтом и бэком.
    Фронт кодируется против контракта (с моками, пока бэка нет), бэк реализует
    контракт. Стыковка = смена базового URL.

    Изменение контракта — только по согласованию обеих сторон. Правило простое:
    **сначала правим contract/openapi.yaml, потом код.**

    ## Ключевое решение: локализация на клиенте

    API **никогда не возвращает переведённый текст**. Возвращаются доменные коды
    (`status: "problem"`, `role: "observer"`) и ключи локализации
    (`name_key: "seg.ustie.name"`). Клиент переводит их через `i18n.js`.

    Причины:
    - фронт уже несёт словарь на 3 языка (ru/en/zh) — дублировать его на бэке не нужно;
    - смена языка не требует запроса к серверу;
    - добавление 4-го языка не затрагивает бэкенд вообще.

    Исключение — **пользовательский контент** (описания в отчётах, имена
    пользователей): возвращается как есть, без перевода.

    ## Что вычисляется только на сервере

    Клиент не может это подделать, поэтому такие значения приходят с бэка и
    никогда не принимаются от фронта:

    - `attention_index` и факторы `s`, `c`, `t`, `a`;
    - переход роли `student → observer` (только после завершения курса);
    - достижение консенсуса по разметке;
...
    - смена статуса участка.
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      security: []
      summary: Регистрация
      description: |
        Возраст определяет режим работы с персональными данными. При `age < 18`
        сервер выставляет `is_minor: true`, закрывает профиль и требует согласие
        родителя для любого офлайн-участия.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, age]
              properties:
                name:     { type: string, minLength: 1, maxLength: 100, example: "Аня" }
                email:    { type: string, format: email }
                password: { type: string, minLength: 8, format: password }
                age:      { type: integer, minimum: 14, maximum: 120, example: 16 }
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409':
          description: Email уже занят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Возраст меньше 14 — регистрация невозможна
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Вход
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/refresh:
    post:
      tags: [auth]
      security: []
      summary: Обновить пару токенов
      description: |
        Access-токен живёт час. Refresh обменивается на новую пару, чтобы
        пользователю не приходилось входить заново в середине разметки.
        Refresh-токен не принимается там, где ожидается access, и наоборот.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Токен истёк, повреждён или неверного типа
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/me:
    get:
      tags: [auth]
      summary: Текущий пользователь
      responses:
        '200':
          description: Профиль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ==================== S2: обучение ====================
...
  /annotations/task:
    get:
      tags: [annotations]
      summary: Получить задание на разметку
      description: |
        Доступно только роли `observer` и выше. Сервер выдаёт участок, который
        пользователь ещё не размечал. Если заданий нет — 204.
      responses:
        '200':
          description: Задание
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnnotationTask' }
        '204': { description: Свободных заданий нет }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Курс не завершён — роль student не допускается к реальной разметке
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /annotations:
    post:
      tags: [annotations]
      summary: Отправить разметку
      description: |
        Координаты — в нормализованных долях `[0..1]` относительно тайла, чтобы не
        зависеть от размера отображения на клиенте.

        Сервер пересчитывает фактор `C` с учётом репутации разметчика и,
        при достижении порога консенсуса, ставит участок в очередь модерации ООПТ.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [segment_id, verdict]
              properties:
                segment_id: { type: string }
                verdict:    { $ref: '#/components/schemas/Verdict' }
                marks:
                  type: array
                  description: Обязательно при verdict != none
                  items: { $ref: '#/components/schemas/Mark' }
      responses:
        '201':
          description: Разметка принята
          content:
            application/json:
              schema:
                type: object
                required: [votes, votes_required, consensus_reached]
                properties:
                  votes:             { type: integer, example: 3 }
                  votes_required:    { type: integer, example: 3 }
                  consensus_reached: { type: boolean }
                  queued_for_moderation: { type: boolean }
                  segment:           { $ref: '#/components/schemas/Segment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Пользователь уже размечал этот участок
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ==================== S4: кабинет ООПТ ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Роль не позволяет выполнить операцию
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message_key]
      properties:
        code:
          type: string
          example: consent_required
          description: Машинный код ошибки
        message_key:
          type: string
          example: err.consent_required
          description: Ключ локализации для i18n.js — текст берётся из словаря фронта
        details:
          type: object
          additionalProperties: true
    Role:
      type: string
      enum: [guest, student, observer, ambassador, oopt_staff, admin]
      description: |
        * `guest` — без регистрации, только тренажёр и публичная карта
        * `student` — зарегистрирован, курс не завершён; **разметка не идёт в очередь ООПТ**
        * `observer` — курс завершён, доступны реальная разметка и акции
        * `ambassador` — активное ядро, может подавать заявки на акции
        * `oopt_staff` — сотрудник ООПТ, только своя территория
        * `admin` — команда проекта
    SegmentStatus:
      type: string
      enum: [problem, work, clean, watch]
      description: |
        * `problem` — выявлена проблема
        * `work` — подтверждён инспектором, в работе
        * `clean` — уборка подтверждена
        * `watch` — под наблюдением
    Verdict:
      type: string
      enum: [dump, litter, none]
      description: dump — скопление отходов, litter — общая замусоренность, none — проблемы нет
    User:
      type: object
      required: [id, name, role, is_minor, reputation]
      properties:
        id:         { type: string, format: uuid }
        name:       { type: string, example: "Аня" }
        role:       { $ref: '#/components/schemas/Role' }
        is_minor:   { type: boolean, description: true при возрасте 14–17 }
        reputation: { type: number, format: float, minimum: 0, maximum: 1.5, example: 1.0 }
        volunteer_hours: { type: number, example: 4 }
        certificate_id:  { type: string, nullable: true, example: "КБ-2026-0417" }
        oopt_id:         { type: string, nullable: true, description: Только для oopt_staff }
        annotations_count: { type: integer, example: 8 }
        accuracy:
          type: number
          format: float
          nullable: true
          description: |
            Измеренная точность разметки по скрытым контрольным заданиям.
            `null`, пока заданий меньше порога: показывать новичку «0%»
            после одной ошибки — способ его потерять. В интерфейсе в этом
            случае выводите «недостаточно данных», а не ноль.
          example: 0.87
        control_tasks_count:
          type: integer
          description: Сколько контрольных заданий пройдено — объём выборки
          example: 40
    AuthResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token: { type: string }
        token_type:   { type: string, enum: [bearer] }
        expires_in:   { type: integer, example: 3600 }
        user:         { $ref: '#/components/schemas/User' }
    AttentionFactors:
      type: object
      description: |
        Факторы индекса внимания. Индекс = 0.40·S + 0.30·C + 0.20·T + 0.10·A.
        Вычисляется сервером, от клиента не принимается.
      required: [s, c, t, a]
      properties:
        s: { type: number, format: float, minimum: 0, maximum: 1, description: Сигнал ДЗЗ }
        c: { type: number, format: float, minimum: 0, maximum: 1, description: Консенсус людей }
        t: { type: number, format: float, minimum: 0, maximum: 1, description: Давность проверки }
        a: { type: number, format: float, minimum: 0, maximum: 1, description: Доступность участка }
    Segment:
      type: object
      required: [id, name_key, status, attention_index, factors]
      properties:
        id:       { type: string, example: ustie }
        name_key:
          type: string
          example: seg.ustie.name
          description: Ключ локализации — фронт переводит через i18n.js
        oopt_id:  { type: string, example: sinie-ozera }
        length_km: { type: number, example: 2.4 }
        status:   { $ref: '#/components/schemas/SegmentStatus' }
        attention_index:
          type: integer
          minimum: 0
          maximum: 100
          example: 77
          description: Готовое к показу значение 0–100
        factors:  { $ref: '#/components/schemas/AttentionFactors' }
        votes:          { type: integer, example: 3 }
        votes_required: { type: integer, example: 3 }
        verified:       { type: boolean }
        last_verified_at: { type: string, format: date-time, nullable: true }
        is_subscribed:  { type: boolean, description: Подписан ли текущий пользователь }
        is_growing:
          type: boolean
          description: |
            Аномалия растёт от снимка к снимку. Маленькая растущая свалка
            требует внимания раньше, чем большая стабильная — показывайте
            это заметным бейджем.
        growth_rate:
          type: number
          format: float
          nullable: true
          description: Прирост площади аномалии между двумя последними сценами, доля
          example: 0.12
...
    Scene:
      type: object
      required: [id, captured_at, source, tile_url_template]
      properties:
        id:          { type: string }
        captured_at: { type: string, format: date }
        source:
          type: string
          enum: [resurs-p, kanopus-v, sentinel-2, uav]
          description: Источник данных — обязателен к показу в интерфейсе
        resolution_m: { type: number, example: 1.0 }
        cloud_cover:  { type: number, minimum: 0, maximum: 1, example: 0.05 }
        tile_url_template:
          type: string
          example: "https://tiles.kosmobereg.ru/scenes/{scene_id}/{z}/{x}/{y}.png"
...
    Mark:
      type: object
      required: [x, y]
      description: Отметка в нормализованных долях тайла [0..1]
      properties:
        x: { type: number, minimum: 0, maximum: 1 }
        y: { type: number, minimum: 0, maximum: 1 }
        r: { type: number, minimum: 0, maximum: 1, description: Радиус зоны, если выделена область }
    AnnotationTask:
      type: object
      required: [segment_id, name_key, scene_before, scene_after]
      properties:
        segment_id:   { type: string }
        name_key:     { type: string }
        scene_before: { $ref: '#/components/schemas/Scene' }
        scene_after:  { $ref: '#/components/schemas/Scene' }
        votes:          { type: integer, description: Сколько разметок уже есть }
        votes_required: { type: integer }
...